requirements.txt
ebooklib
google-genai
click
beautifulsoup4
ollama